Update Bucket policy (#277)
Modified the resource S3DataBucketPolicy, Line 342, PolicyDocument. Added "!Sub "${S3Bucket.Arn}"" to the Resource in both statements, in order to apply the policy to the bucket itself and all objects within it. 

For example, this change is required in order to pass a  Control Tower proactive control (CloudFormation Hook):
[CT.S3.PR.8] Require that Amazon S3 bucket requests use Secure Socket Layer

https://docs.aws.amazon.com/controltower/latest/controlreference/s3-rules.html#ct-s3-pr-8-description